ADW app with AEV/assets/source-code/adw-3.1.1-aev/libs/content-ee/process-services-cloud-extension/src/lib/features/task-details/components/task-assignment-dialog/task-assignment-dialog.component.ts
/*
 * Copyright © 2005 - 2021 Alfresco Software, Ltd. All rights reserved.
 *
 * License rights for this program may be obtained from Alfresco Software, Ltd.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

import { Component, OnInit, Inject } from '@angular/core';
import { MAT_DIALOG_DATA } from '@angular/material/dialog';
import { TaskAssignmentService } from '../../services/task-assignment.service';
import { UntypedFormControl } from '@angular/forms';
import { TaskAssigneeModel } from '../../models/task-assignee.model';
import { IdentityUserModel, IDENTITY_USER_SERVICE_TOKEN } from '@alfresco/adf-process-services-cloud';

@Component({
    selector: 'apa-candidates-dialog',
    templateUrl: './task-assignment-dialog.component.html',
    providers: [TaskAssignmentService, { provide: IDENTITY_USER_SERVICE_TOKEN, useExisting: TaskAssignmentService }],
})
export class TaskAssignmentDialogComponent implements OnInit {
    isCurrentAssigneeSelected = true;
    preselectedAssignee: IdentityUserModel[] = [];
    searchUserControl = new UntypedFormControl('');
    selectedAssignee: IdentityUserModel;

    constructor(
        private taskAssignmentService: TaskAssignmentService,
        @Inject(MAT_DIALOG_DATA) public settings: TaskAssigneeModel) {}

    ngOnInit() {
        if (this.settings.assignee) {
            this.preselectedAssignee = [{ username: this.settings.assignee }];
        }
        this.taskAssignmentService.setApplicationName(this.settings.appName);
        this.taskAssignmentService.setTaskId(this.settings.taskId);
    }

    onSelect(assignee: IdentityUserModel) {
        if (assignee.username.toLocaleLowerCase() !== this.settings.assignee.toLocaleLowerCase()) {
            this.selectedAssignee = assignee;
            this.isCurrentAssigneeSelected = false;
        } else {
            this.isCurrentAssigneeSelected = true;
        }
    }

    onRemove() {
        this.selectedAssignee = null;
    }

    isAssignButtonDisabled(): boolean {
        return this.isCurrentAssigneeSelected || this.searchUserControl.invalid || this.selectedAssignee === null;
    }
}
